+++ modules/HyperIoTHProject/HyperIoTHProject-model/src/main/java/it/acsoftware/hyperiot/area/model/Area.java
@@ -95,6 +95,11 @@ public class Area extends HyperIoTAbstractEntity
     @JsonView({HyperIoTJSONView.Extended.class, HProjectJSONView.Export.class})
     private Set<Area> innerArea;
 
+    public Area() {
+        //default view type is image
+        this.areaViewType = AreaViewType.IMAGE;
+    }
+
     /**
      * @return Area name
      */